requirements.txt
python-socketio
flask
eventlet